hotfix(no-ticket): update quantity available schema

+++ src/modules/api/products/common/schema/product-quantity-available.schema.ts
@@ -1,3 +1,3 @@
 import { z } from 'zod';
 
-export const productQuantityAvailableSchema = z.number().min(0);
+export const productQuantityAvailableSchema = z.number().int().min(0);
